Fixed a few bugs, import errors and changed to asyncio wait_for instead of timeout to support python versions < 3.11
